fix: Split UrlError into separate file, expose to test target
